// src/generated/geo-autocomplete.ts
export interface components {
    schemas: {
        /** @description Contains location predictions for the text-based geographic search query. */
        GeoAutocompleteResponse: {
            /** @description The metadata associated with the request and response. */
            responseHeader?: {
                /**
                 * @description HTTP status code.
                 * @example 0
                 */
                status?: number;
                /**
                 * @description The time (in seconds) taken to execute the query.
                 * @example 4
                 */
                QTime?: number;
            };
            /** @description The results of the query specific to the requested category. */
            grouped?: {
                "category:AIR"?: components["schemas"]["CategoryResponse"];
                "category:RAIL"?: components["schemas"]["CategoryResponse"];
                "category:CITY"?: components["schemas"]["CategoryResponse"];
                "category:POI"?: components["schemas"]["CategoryResponse"];
                "categpry:LOCATION"?: components["schemas"]["CategoryResponse"];
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}